requirements.txt
# Deep Learning
torch>=2.0.0

# Data Processing
numpy>=1.24.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Interface
streamlit>=1.20.0

# Evaluation Metrics
scikit-learn>=1.2.0

# AWS Simulation (for future cloud integration)
boto3>=1.26.0
moto>=4.0.0